requirement.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==2.3.3
Flask-Cors==4.0.0
google-auth==2.45.0
google-genai==0.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.0
opencv-python==4.8.1.78
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.2.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==14.2
Werkzeug==2.3.7
